busybox: Remove old ipv6 addresses in bound step
In upstream Busybox the udhcpc and udhcpc6 clients generate a deconfig
event when sent certain signals, and on exit.

The udhcp script needs to know which client (v4 or v6) has called it in
order to know which address type to clear from the interface, however
there is not current a way of determining this. It can't do both, or a
sequence of

 deconfig (ipv6), bound (ipv6), deconfig (ipv4)

will lose the ipv6 address.

This updates the script to check for ipv6 addresses that were previously
bound by udhcpc6 and remove them before adding the new address.

Without this a system will accumulate IPv6 addresses over time if it
re-queries the network periodically.

Signed-off-by: Samuel Mendoza-Jonas <sam@mendozajonas.com>